requirements.txt
google-adk[mcp]
requests
